+++ packages/section-player/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @pie-players/pie-section-player
 
+## 0.3.45
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+- Updated dependencies [fd140a3]
+  - @pie-players/pie-assessment-toolkit@0.3.45
+  - @pie-players/pie-context@0.3.45
+  - @pie-players/pie-item-player@0.3.45
+  - @pie-players/pie-players-shared@0.3.45
+  - @pie-players/pie-tool-annotation-toolbar@0.3.45
+  - @pie-players/pie-tool-calculator-desmos@0.3.45
+  - @pie-players/pie-tool-text-to-speech@0.3.45
+  - @pie-players/pie-tool-tts-inline@0.3.45
+  - @pie-players/pie-toolbars@0.3.45
+
 ## 0.3.44
 
 ### Patch Changes